Refactor extra info
